Add pasting of text into the remote machine.
Two stub functions to handle sensitive (passwords etc.) and
non-sensitive inputs. They, in turn, call a generic function which
essentially emulates a keypress for each character.

Things get a bit hairy when dealing with characters produced with a
modifier: uppercase is fairly straightforward, but symbols seem to be
dependent on keyboard settings on the host.

A lookup table to convert from character to keycode is needed since
usually the keycode comes for free from the keypress event.