Quick sketch of the intended solution
We should make the paste button look a little prettier.
There seems to be a sync issue where too many cards are created and they go out of sync with the success/fail colors.
The overlay should be more readable and we should make sure normal ways to paste work.
We should also make it actually honor the 'any other key to cancel' and not pass it through to the target system.