Add more rounded paste functionality.
The overlay now places the cursor at the end of the text, so you can
immediately hit Ctrl/Cmd + V and it'll paste into the remote machine.

Any other key will close the window immediately, with the exception of
Ctrl since if you're a bit slow to hit Ctrl/Cmd + V (like me, it seems)
then the overlay closes before you've completed the key combination.
This is a purely cosmetic allowance.